+++ src/pages/Register.js
@@ -1,5 +1,4 @@
 import React from 'react'
-import { Form, Input, Button, Checkbox } from 'antd';
 import { useState } from "react";
 import { useDispatch } from "react-redux";
 import { useHistory } from "react-router-dom";
