--- src/redux/realestate/realestateTypes.js
@@ -1,2 +1,0 @@
-export const REALESTATE_SUCCESS = 'REALESTATE_SUCCESS';
-export const REALESTATE_FAILURE = 'REALESTATE_FAILURE';
