+++ src/redux/realestates/realestatesReducer.js
@@ -1,9 +1,13 @@
-import React from 'react';
-
-import { REALESTATES_SUCCESS, REALESTATES_FAILURE } from './realestatesTypes'
+import { 
+  REALESTATES_SUCCESS, 
+  REALESTATES_FAILURE,
+  REALESTATE_SUCCESS,
+  REALESTATE_FAILURE
+ } from './realestatesTypes'
 
 const initialState = {
   realestates: null,
+  realestate: null,
   errors: ''
 };
 
@@ -15,6 +19,16 @@ const realestatesReducer = (state = initialState, action) => {
         realestates: action.payload
       }
     case REALESTATES_FAILURE:
+      return {
+        ...state,
+        errors: action.payload
+      }
+    case REALESTATE_SUCCESS:
+      return {
+        ...state,
+        realestate: action.payload
+      }
+    case REALESTATE_FAILURE:
       return {
         ...state,
         errors: action.payload
