+++ src/redux/realestates/realestatesTypes.js
@@ -1,2 +1,5 @@
 export const REALESTATES_SUCCESS = 'REALESTATES_SUCCESS';
 export const REALESTATES_FAILURE = 'REALESTATES_FAILURE';
+
+export const REALESTATE_SUCCESS = 'REALESTATE_SUCCESS';
+export const REALESTATE_FAILURE = 'REALESTATE_FAILURE';
